Refactored structure of bot package
